Separate `URLMatcher` from `DebounceService`
